requirements.txt
psycopg2-binary>=2.9.9
pgvector>=0.3.0
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.26.0
gspread>=6.0.0
google-auth>=2.28.0
voyageai>=0.3.0
sentence-transformers>=3.0

# Chatbot server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=2.0.0
huggingface-hub>=0.29.0
pydantic>=2.6.0

# Voice (Twilio)
twilio>=9.2.3
